refactor(table): extract transaction validation into its own module
validate_operation and its four schema/fragment helpers check an operation
against the manifest it applies to. They have no other callers inside
transaction.rs beyond one call to merge_fragment_physically_rewritten from
build_manifest.

Also adds transaction::test_support for the three fixtures that more than one
submodule's tests will need, so later extractions have somewhere to reach for
them.
